Written by: Edd Barrett <edd@openbsd.org>
Found bogus junk being installed in ${PREFIX}/texmf*. Also showed me
that some symlinks need to be translated. Unfortunately this now causes
symlink warnings during packaging; this is because teh relative symlinks
are bogus in the context of ${WRKINST}, but not once relocated at
install time. Any ideas what to do with that anyone?

Not linked to build, no OK.