Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Move to MODPY_ADJ_FILES.

There is no point right now to dig the entire tree to move ports to using
this, but for new ports and when updating old ones, please try to use it
when possible. I just committed this one as an exemple on how to use this.