Written by: Jeremy Evans <jeremy@openbsd.org>
Update for the new ruby.port.mk to add the ruby18 WANTLIB back. I've
already prepared an update for this, but it hasn't been approved yet,
and this port needs to be working in the meantime.  Since it's
possible someone could have built the port since the ruby.port.mk
change, bump for safety.

ok naddy