Written by: Landry Breuil <landry@openbsd.org>
Add a bunch of patches from Todd Carson <toc at daybefore dot net> to
allow webkit to build (and even run!) on mips64el, many thanks !
Patches will probably need to be reworked so that they can be fed
upstream, but at least this allows another part of the tree to build on
loongson. And this might even work for sgi.
Tested by jasper@