Written by: Robert Nagy <robert@openbsd.org>
Initial import of LibreOffice 3.3rc3;

LibreOffice is a productivity suite that is
compatible with other major office suites,
and available on a variety of platforms.
It is free software and therefore free to
download, use and distribute. 

The port is not yet ready and not linked
to the build yet, but it builds and works
fine. All of the OpenBSD specific patches
have been commited to upstream already.