Written by: Jeremy Evans <jeremy@openbsd.org>
Add a PKGSPEC entry to restrict ports depending on lang/ruby/1.9 to
depend on ruby-1.9 and not ruby-1.8.  This PKGSPEC is slightly
different from the previous one used in ruby.port.mk, so all
dependent ports need to be REVISION bumped (which will happen soon).
In addition, since the subpackages depend on the -main package and
also had a slightly different pkgspec, they need to be bumped as well.

In addition, the -main package is also being bumped for the inclusion
of a patch for Addrinfo that fixes a failing IPv6 UDP regression test.

OK landry@