Written by: Jeremy Evans <jeremy@openbsd.org>
Install the binaries with the same names as upstream, fixes building
ruby gem ports with the jruby FLAVOR in some cases.

Don't install the rdoc for the rake and rspec gems, for better
consistency with other ruby gem ports.

Add a @comment file to PLIST to fix update-plist adding a file that
shouldn't be in the package.

OK landry@