Written by: Jeremy Evans <jeremy@openbsd.org>
Handle the gem metadata such that if a build fails and you restart
it, it doesn't fail again just because it can't find the gem
metadata.

Also, get rid of the GEM_REL variable, and just setup the GEM_BIN
and GEM_LIB variables separately for each ruby version.

OK landry@