Written by: Jeremy Evans <jeremy@openbsd.org>
Import ruby-arel, one of the dependencies for Rails 3.

Arel is a Relational Algebra for Ruby. It 1) simplifies the generation
complex of SQL queries and it 2) adapts to various RDBMS systems. It is
intended to be a framework framework; that is, you can build your own
ORM with it, focusing on innovative object and collection modeling as
opposed to database compatibility and query generation.

ok landry