Written by: Landry Breuil <landry@openbsd.org>
Import chroma 1.13:

Chroma is an abstract puzzle game. A variety of colourful shapes are arranged
in a series of increasingly complex patterns, forming fiendish traps that must
be disarmed and mysterious puzzles that must be manipulated in order to give up
their subtle secrets. Initially so straightforward that anyone can pick it up
and begin to play, yet gradually becoming difficult enough to tax even the
brightest of minds.

From Daniel Levai on ports@ with tweaks by me, ok jasper@