Written by: Landry Breuil <landry@openbsd.org>
Import p5-Filesys-Notify-Simple:

Filesys::Notify::Simple is a simple but unified interface to get
notifications of changes to a given filesystem path. It utilizes
inotify2 on Linux and fsevents on OS X if they're installed, with a
fallback to the full directory scan if they're not available.

(note from importer: it could use a kqueue/libevent backend instead of a
gross full scan every 2 secs, which is not even configurable :)

From Abel Abraham Camarillo Ojeda on ports@