Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
There seems to be an annoying runtime regression with session bus after
the last update which I failed to see in my tests.
While this is being worked on, let's revert for now so that people don't
run into it. Other parts of the D-DBus gang update are fine (glib and
python). Also keep the recent rc script tweak.

Sorry about that...