Written by: Stephan A. Rickauer <stephan@openbsd.org>
import of Metasploit 3.5.0, the open source penetration testing framework:

Metasploit provides useful information and tools for penetration
testers, security researchers, and IDS signature developers. This
project was created to provide information on exploit techniques and to
create a functional knowledgebase for exploit developers and security
professionals. The tools and information on this site are provided for
legal security research and testing purposes only.

Subpackages:
        mysql - mysql database backend.
        postgresql - postgresql database backend.

help from nicm@, benoit@ and jeremy@
ok jasper@ and jeremy@ (of some earlier version)