Written by: Matthias Kilian <kili@openbsd.org>
This package contains the generics system described in the `Scrap
Your Boilerplate' papers (http://www.cs.uu.nl/wiki/GenericProgramming/SYB).
It defines the Data class of types permitting folding and unfolding
of constructor applications, instances of this class for primitive
types, and a variety of traversals.

Needed by some ports after an update to ghc-7.0.2, not yet linked to the build.

Reads good to landry@