Written by: Matthias Kilian <kili@openbsd.org>
Import mtl-2.0.1.0, needed by a couple of hs-* ports after the
update to ghc-7.0.2. Unfortunately, there are some other ports
that don't build with mtl-2 yet, so in the worst case i've to
include mtl-1 within the ghc package and mtl-2 in this new port.

Not yet linked to the build, of course.