Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import obc-2.9.1

Oberon-2 is clean and simple programming language.
This is portable compiler that translates Oberon-2 into bytecode, which
can be either interpreted or dynamically translated into machine code.
Dynamic translation uses a portable interface modelled on the one in GNU
Lightning, but is so far implemented only on x86 machines.  The
implementation includes a full garbage collector, and comes with
profiling tools and a simple GUI debugger.
[...]

    from Alexander Shiryaev (MAINTAINER) with tweaks by me
	
	    ok landry@