Written by: Landry Breuil <landry@openbsd.org>
Update to libxfcegui4 4.8.0.

The library is still needed by some Xfce apps and most panel plugins,
but porting to the new libxfce4ui is on the way. It doesn't ship anymore
libxfce4kbd-private which moved to libxfce4ui, but the name is not
exactly the same so they don't conflict.