Written by: Edd Barrett <edd@openbsd.org>
Update to TeX Live 2010.

This year we pre-generate format and maps as requested by kili.

Started adding test cases in the tests/ dir.

Upgrades from 2009 will cause a harmless warning regarding texlinks. Espie is
aware of this, it was not easily fixable.

Uninstalls of texlive 2010 which were previously upgraded from 2009 will make
lots of noise regarding non-empty directories which can not be deleted. This
is due to unregistered package directories in the old texlive 2009 pacakage and
has been fixed in the 2010 package (rblatter script was modified to register
*all* parent dirs in PLISTs).

Installs/uninstalls of texlive 2010 on a fresh system (which has never
seen texlive 2009) are squeaky clean.

This has been in use for a number of months on my work machine, so should be
pretty stable, but I am sure someone will find some obscure bugs. These can be
fixed in tree.

OK espie@