Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import mm-common-0.9.3.

The mm-common is part of the GNOME C++ bindings effort.
It provides the build infrastructure and utilities shared among the
GNOME C++ binding libraries.  It is only a required dependency for
building the C++ bindings from the gnome.org version control repository.
An installation of mm-common is not required for building tarball
releases, unless configured to use maintainer-mode.

ok jasper@