Written by: Jeremy Evans <jeremy@openbsd.org>
Add MODRUBY_HANDLE_FLAVORS variable which can be set to No to not have
the lang/ruby module add FLAVORS automatically for gem and extconf
ports, or to Yes to enable the FLAVORS for non-gem and non-extconf
ports.

Add a MODRUBY_BIN_TESTRB variable containing the appropriate testrb
command for use in regress tests.

OK landry@