Written by: Michael Erdely <merdely@openbsd.org>
From Pascal Stumpf with input from ajacotout@:
"this is a small update of privoxy to 3.0.17, plus an rc script making
use of the built-in chroot functionality. For this to work, you need to
change the home of _privoxy to /var/privoxy, the rc script will copy all
necessary files into the chroot.

Also, some additional files and scripts will be installed."

Similar diff and testing from roberth
Input and OK from sthen@