Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import nagstamon-0.9.4.

Nagstamon is a Nagios status monitor for the desktop. It connects
to multiple Nagios, Icinga, Opsview and Centreon monitoring servers
and resides in systray or as a floating statusbar at the desktop
showing a brief summary of critical, warning, unknown, unreachable
and down hosts and services and pops up a detailed status overview
when moving the mouse pointer over it. Connecting to displayed hosts
and services is easily established by context menu via SSH, RDP and
VNC. Users can be notified by sound. Hosts and services can be
filtered by category and regular expressions.

Note that systray support (as opposed to the floating status bar)
requires the "py-gnome-extras" package. Dependency is not enforced on
OpenBSD as this would bring lots of other dependencies.

ok sthen@