Written by: Robert Nagy <robert@openbsd.org>
make sure that we don't try to use NPTL code because if we do, all of the
zarafa-spooler processes will end up as zombies and we won't be able
to send emails.
this commit also removed the upper limit of 1 thread and from now on
we use the default value which is 5.

ok ajacoutot@