Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import nakeasn 20110313

This is an open-sourced assembler / disassembler / simulator for the
MSP430 series of microcontrollers from Texas Instruments. The goal of
this project is to have a completely cross-platform assembler with the
exe (naken430asm) under 50k and to have a cross-platform disassembly
utility (naken430util) along with it to help with debugging and a bit of
profiling. The naken430util, when disassembling, shows how many CPU
cycles each instruction will take and it has a built in simulation to help
show how the binary code will flow on the chip and how many cycles it
will take to execute.
Code from this assembler can be uploaded to the device by installing the
mspdebug package.

ok aja@