Written by: Stuart Henderson <sthen@openbsd.org>
import pear-Date-Holidays, from Johan Huldtgren with tweaks by myself
and jasper@, ok jasper.

Date_Holidays helps you calculating the dates and titles of holidays
and other special celebrations. The calculation is driver-based so
it is easy to add new drivers that calculate a country's holidays.
The methods of the class can be used to get a holiday's date and
title in various languages.