Written by: Landry Breuil <landry@openbsd.org>
Import fennec 4.0:

Firefox for mobile (codenamed Fennec) is the name of the build of the
Mozilla Firefox web browser for devices such as mobile phones and
personal digital assistants (PDAs).
While it initially targets android and maemo, it also works on desktop
machines, and can be useful on small screens or low-powered devices.

Looks ok to jasper@