Written by: Stuart Henderson <sthen@openbsd.org>
Import ports/productivity/radicale, a simple CalDAV server.
Tweaked from a submission from maintainer Sergey Bronnikov, ok landry@

The Radicale Project is a complete CalDAV calendar server solution,
capable of making multiple calendars available to local and remote
users, with optional authentication policies. Calendars can be
viewed and edited by a calendar client such as Sunbird or Evolution.

The Radicale Project aims to be a light solution, easy to use, easy
to install, easy to configure. As a consequence, it requires few
software dependencies and is pre-configured to work out-of-the-box.