Written by: Stuart Henderson <sthen@openbsd.org>
Have check_mailscanner pick up the "Run As User" setting from
MailScanner.conf and explicitly start as that user. This avoids setuid
in perl (triggering taint mode) and so it sidesteps taint mode fallout
which has appeared recently, most likely due to the CVE-2011-1487 fix
in Perl.

Maintainer believed to be away this week, so committing now and we
can change it later if he disagrees when he returns.   ok todd@