Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Workaround the fact that we cannot pass SCM_CREDS over unix sockets:
implement the "intended" SCM_CREDS stack as if we had support for that
in the kernel (by-pass it almost completely).
send/recv a single null byte without creds, but on recv, just do a
getsockopt(SO_PEERCRED) and return that as if it coming from the cmsg.
This works as long as creds are not retreived from an fd which has
already been handed over to a different process via SCM_RIGHTS. It will
probably not be enough in the future but we'll see then.
all this work done by eric@ (thanks!) and tested by myself

Enable support for g_credential*
Fix a couple of warnings.

ok eric@ jasper@