Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import gnote 0.7.4

Gnote is a port of Tomboy to C++.
It is the same note taking application, including most of the add-ins
(more are to come). Synchronization support is being worked on.

This will replace tomboy in our tree, users can sync their tomboy notes
with the TomboyImport addin.

feedback/ok aja@