Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Add BUILD_DEPENDS on rarian. Technically it is not entirely needed but
since we depend on yelp it'll get installed anyway. This is to ease
grep(1) over the ports tree to check ports that could be impacted by
the GNOME3 update (as well as consistency with the rest of stuffs
dependeing on yelp).

While here, regen wantlib.