Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Add an x11/gnome3 subdirectory.
This is a staging area that will not be hooked into the build. It's used
to commit updates to x11/gnome that cannot yet make it in the regular
tree before we move to GNOME3. When the time is come, everything under
x11/gnome3 will be moved back to x11/gnome.

Using this will allow us to have everything in-tree and allow brave
souls to play, help, test... :-)

discussed with and ok jasper@