Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Shuffle things around so that the gnome3 directory is usable like
mystuff. README coming soon so anyone can start having 'fun' with gnome3
ports.
Also this will allow proper dependencies to be set from the start
instead of breaking half of the tree when GNOME3 is enabled.

discussed with an ok jasper@