Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Add gtk+2 and librsvg to RUN_DEPENDS.

Rational: except for QT/KDE ports, gtk+2 should already be in the
dependency chain for ports installing share/icons/ files (i.e. no need
to explicitely add it) and if we have svg icons without their png
counterpart and librsvg is not part of the dependency chain then we
explicitely add it.
gnome-icon-theme is the inherited theme for most icon themes so we
make sure we have gtk-update-icon-cache and can display svg icons.

landry@ agrees