Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import metacity-2.34.0.

Metacity is a simple window manager that integrates nicely with GNOME.

(note that it is exactly the same as the one under x11/gnome except it
depends on x11/gnome/themes-standard (i.e. gtk+3) because the default
theme is different in GNOME3 -- eventually when upstream fix gtk+3
support in metacity this can go back in the main hierarchy and we'll add
a gtk+3 FLAVOR)