Written by: Robert Nagy <robert@openbsd.org>
Add WIP progress sndio support to chrome, developed by jakemsr@ and
with modifications by me.
It is not working yet so you can only enable it by setting the
CHROME_SNDIO environment variable, so by default nothing changes.
I commit this so that other people can work on it too.