Written by: Stuart Henderson <sthen@openbsd.org>
iplayer-dl used the BBC's iPhone streams to download programmes; however
BBC changed things in Dec 2010 to require a TLS connection with an Apple-
signed client certificate present in Apple iOS devices. As a result
iplayer-dl is no longer useful, so remove the port.

(get_iplayer is an alternative which uses the RTMP streams intended
for the Flash player instead).