Written by: Landry Breuil <landry@openbsd.org>
Import/update to warmux 11.04.1, from Daniel Levai.

This is the same port as games/wormux and will update it, as upstream
was renamed.

Have the mascots of your favorite free software battle in the Warmux
arena. Using dynamite, grenades, baseball bat and others bazookas,...
exterminate your opponent in a 2D toon style scenery and a funny
environment.