Written by: Matthias Kilian <kili@openbsd.org>
Zaz is a game where the player has to get rid of incoming balls by
arranging them in triplets. The idea of the game is loosely based
on games like Luxor, Zuma and Puzzle Bobble. The twists that make
Zaz stand out from other games of this type are that the balls have
to be picked from the path (insted of being randomly assigned for
the player) and that the player's "vehicle" is also attached to a
path which is different from level to level.

ok landry@ (after removing USE_X11 and after jasper fixed ftgl)