Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import tryton{,d}-2.0.1

Tryton is a three-tiers high-level general purpose application platform
written in Python and uses Postgresql as its main database engine. It is
the core base of an Open Source ERP. It provides modularity, scalability
and security.

"tryton" is the client of the Tryton application platform.
"trytond" is the server of the Tryton application platform.