Written by: Marc Espie <espie@openbsd.org>
fix some limitations of _print-package-args
* cut it into separate targets for readability and better testing
* filter libraries more efficiently
* use internal variable for resolve-lib

create two targets for solving wantlib: one (_fake-wantlib-args) which uses
the information under the fake directory and installed packages, and another
(_port-wantlib-args) which walks packing-lists. The second one can be used
to collect meta-info even when nothing is installed, and thus provide better
package-signature accuracy, or help with lib-depends-check.
The first one uses what's actually in place when a package is built.
For now, we're paranoid and use both when building a package, erroring out
if they don't match, even though _port-wantlib-args is somewhat slower.