Written by: Stuart Henderson <sthen@openbsd.org>
vim's tar support requires GNU tar to operate correctly; patch the relevant
script to use "gtar" not "tar". This is not added as a run dep to avoid
pulling in extra dependencies, especially xz which requires gcc3+ so is
not available on some arch and is more difficult to build on others.
(There's a clear error message at runtime if gtar is unavailable.)
ok dcoppa@ ckuethe@