Written by: Stuart Henderson <sthen@openbsd.org>
revert last and add xvidcore to WANTLIB instead, allowing the
library dependency to be tracked properly. even though it's
dlopen(3)'d not linked directly, we still need to make sure the
package is compatible with the library.

it will show as Extra in lib-depends-check, that's not important
(but add a comment to make sure people don't think it can be removed).

pointed out by jakemsr and discussed with brad and espie.