Written by: Stuart Henderson <sthen@openbsd.org>
move WANTLIB+=${MODPY_WANTLIB} within the .if ${FLAVOR:L:Mpython} block;
no package change as the python module is only included if the python
flavour is selected (and unless that's included, MODPY_WANTLIB is blank),
but it's clearer this way.

side-note, python bindings should probably be split into a subpackage
sometime, there seems to be no need to have a separate flavour for this.