Written by: Stuart Henderson <sthen@openbsd.org>
Handle the dlopen()'d libavcodec the same way as transcode does with xvidcore;
move it from a BUILD+RUN_DEPENDS to WANTLIB+LIB_DEPENDS so that the dependency
is tracked and opal gets updated if the FFmpeg API changes. Comment to explain
why check-lib-depends says it's extra. Pointed out by Brad, ajacoutot@ hates
it but is ok with it.