Written by: Stuart Henderson <sthen@openbsd.org>
Remove patch which was removing various -W flags. The updated timestamp on
the file triggers an attempt to rebuild autoconf files, which breaks the build
if automake-1.11 is installed.  Worked out by Nigel Taylor.

Fix typo in DESCR while there.  The patch was there because these flags
broke the build in an earlier version of libnice, but are no longer needed
(checked gcc 3 and 4; port didn't build on gcc 2 arch anyway).  ok jasper@