Written by: Stuart Henderson <sthen@openbsd.org>
import ports/net/scamper, ok benoit@

scamper is a program that is able to conduct Internet measurement
tasks to large numbers of IPv4 and IPv6 addresses, in parallel, to
fill a specified packets-per-second rate. Currently, it supports
the well-known ping and traceroute techniques, as well as radargun,
ally and mercator. (On some other OS it also supports sting, and
parts of tbit).