Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import sslscan-1.8.2.

SSLScan queries SSL services, such as HTTPS, in order to determine the
ciphers that are supported. SSLScan is designed to be easy, lean and
fast. The output includes prefered ciphers of the SSL service, the
certificate and is in Text and XML formats.

Heavily based on the FreeBSD port.

inputs from sthen@
ok jasper@