Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import trytond-module-product_cost_fifo-2.0.1.

This is the "product_cost_fifo" module for the Tryton application
platform.
It adds FIFO cost method on the product form.
Once installed, the price of the stock moves from a supplier or to a
customer will update automatically the cost price of the related
product (if the cost price method is FIFO).

from Christian Rueger with some tweaks
ok jasper@